Clean environment prior to starting and wait properly for orchestrator to stop
The current init script does not clean up the environment so this
may make orchestrator behave different depending on the environment
of the caller.

When stopping orchestrator sometimes takes time to shutdown, and
if you do not wait starting it will fail as it is likely to be still
bound to the listen port. This change ensures we wait for orchestrator
to terminate properly.