Optimise scanning binlogs when using pseudo-gtid
When moving a server around the topology the latest pseudo-gtid
may be needed. To get this orchestrator must scan the last binlog
until the end to grab this latest event.

This change compiles the regexp used to search for psuedo-gtid
events and uses the compiled version.

On a 100 MB binlog with ~ 1.3 million binlog events it reduces the
time from ~160 seconds to ~10 seconds.