backend write optimization: bulk instance writes
Instead of updating db on every instance we accumulate
writes on instanceWriteBufferSize queue and
flush it to the db periodically.

This commit generalizes insert odku query construction
as well.

At the moment, we flush the queue as soon as it grows
to 100 items. That's good enough to experiment with
but definitely not good enough for every environment.
We will change this to a timer-base flushes later.
To be continued.