RecoverDeadMaster: Add validation of failedInstanceKey
The following has been seen:

... topology_recovery: RecoverDeadMaster: will recover :0

While this should not happen, and may be due to other non-outbrain
code, the result can be quite damaging given the way that orchestrator
then looks for all slaves replicating from this master and decides
to reorganise them.

This added check should prevent this situation happening and is an
extra and cheap safety net.