move global recoveries check down one func call
Because tons of checks happen at the top of executeCheckAndRecoverFunction,
that return from the function without triggering any recoveries.
By moving the check we avoid noise on non-recoveries (which is
happening quite a lot).

This also makes global recoveries flag properly act on
    specificInstance != nil

Additionally, some extra logging is added on bailing out
of executeCheckAndRecoverFunction.