Fix some problems with pseudo-GTID match of a 5.7 slave under a 5.6 master.
Analysis:

When upgrading major versions you may run two different major
versions in the same replication chain. You may also have intermediate
masters of one version or another. Slaves should be at least as
"new" as their upstream master.

With a 5.6 master, replicating through a 5.7 intermediate master
to a 5.7 slave this process will not work properly if you try to
relocate the slave under a 5.6 master again. The reason for this
is that 5.7 when not in GTID mode emits "anonymous GTID next" events
which are not in the binlog stream of the 5.6 master. End result
is a broken matching process.

Initial attempts to resolve this were to use the setting

 "SkipBinlogEventsContaining": [
    "SET @@SESSION.GTID_NEXT= 'ANONYMOUS'"
  ],

but this has its own issues as:
(a) it is applied on all systems and not where strictly speaking
it is needed and
(b) it also "hides" some events from the matching logic triggering
a failure to relocate if the slave is stopped on an anonymous GTID_NEXT
event..

The new logic recognises the specific case and only ignores these
anonymous events on the binlog/relaylog stream of the server being
relocated.